added test fix

+++ src/Layers/W1/Tests/User/TestUserTasks.Codeunit.al
@@ -188,26 +188,29 @@ codeunit 134769 "Test User Tasks"
         FieldRec.SetFilter(Type, '%1|%2', FieldRec.Type::Code, FieldRec.Type::Text);
         if FieldRec.FindSet() then
             repeat
-                TableInformation.SetFilter("Company Name", '%1|%2', '', Company.Name);
-                TableInformation.SetRange("Table No.", FieldRec.TableNo);
-                if TableInformation.FindFirst() then begin
-                    RecRef.Open(FieldRec.TableNo, false, Company.Name);
-                    if TempTablesAlreadyInserted.Get(FieldRec.TableNo) then begin
-                        RecRef.FindFirst();
-                        FldRef := RecRef.Field(FieldRec."No.");
-                        FldRef.Value('OLD');
-                        RecRef.Modify();
-                    end else begin
-                        RecRef.DeleteAll();
-                        RecRef.Init();
-                        FldRef := RecRef.Field(FieldRec."No.");
-                        FldRef.Value('OLD');
-                        RecRef.Insert();
-                        TempTablesAlreadyInserted.Init();
-                        TempTablesAlreadyInserted.Number := FieldRec.TableNo;
-                        TempTablesAlreadyInserted.Insert();
+                if not SkipTable(FieldRec.TableNo) then begin
+                    TableInformation.SetFilter("Company Name", '%1|%2', '', Company.Name);
+                    TableInformation.SetRange("Table No.", FieldRec.TableNo);
+                    if TableInformation.FindFirst() then begin
+                        RecRef.Open(FieldRec.TableNo, false, Company.Name);
+                        if TempTablesAlreadyInserted.Get(FieldRec.TableNo) then begin
+                            RecRef.FindFirst();
+                            FldRef := RecRef.Field(FieldRec."No.");
+                            FldRef.Value('OLD');
+                            RecRef.Modify();
+                        end else begin
+                            RecRef.DeleteAll();
+                            RecRef.Init();
+                            FldRef := RecRef.Field(FieldRec."No.");
+                            FldRef.Value('OLD');
+                            RecRef.Insert();
+
+                            TempTablesAlreadyInserted.Init();
+                            TempTablesAlreadyInserted.Number := FieldRec.TableNo;
+                            TempTablesAlreadyInserted.Insert();
+                        end;
+                        RecRef.Close();
                     end;
-                    RecRef.Close();
                 end;
             until FieldRec.Next() = 0;
 
@@ -233,6 +236,8 @@ codeunit 134769 "Test User Tasks"
     local procedure SkipTable(TableNo: Integer): Boolean
     begin
         if TableNo = DATABASE::"Fin. Report Package Recipient" then
+            exit(true);
+        if TableNo = DATABASE::"Contact Sync User" then
             exit(true);
         exit(false);
     end;
